# shared/contracts/schemas/openapi.yaml
openapi: "3.1.0"
info:
  title: Lychee Ripe Gateway API
  version: "1.1.0"
  description: >
    External API served by the Go gateway. The gateway provides auth/rate-limit,
    forwards inference requests, and exposes business APIs for batch traceability.

servers:
  - url: http://localhost:9000
    description: Local gateway

paths:
  /healthz:
    get:
      operationId: gatewayHealth
      summary: Aggregated gateway and upstream health
      tags: [system]
      security: []
      responses:
        "200":
          description: All systems healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayHealthResponse"
        "503":
          description: Upstream degraded or unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayHealthResponse"

  /v1/health:
    get:
      operationId: upstreamHealth
      summary: Upstream FastAPI health (proxied)
      tags: [v1]
      security: []
      responses:
        "200":
          description: Upstream health
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /v1/models/current:
    get:
      operationId: currentModel
      summary: Current loaded model info (proxied)
      tags: [v1]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        "200":
          description: Model metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentModelResponse"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden by policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Upstream unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/infer/image:
    post:
      operationId: inferImage
      summary: Single image inference (proxied)
      tags: [v1]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Inference result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageInferResponse"
        "400":
          description: Invalid image or request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden by policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Model not loaded or upstream unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/infer/stream:
    get:
      operationId: inferStream
      summary: WebSocket streaming inference (proxied)
      tags: [v1]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      description: >
        WebSocket endpoint. Send binary frames (image bytes) to receive per-frame
        detection results. Send text "close"/"stop"/"eos" to end. On disconnect,
        server sends a SessionSummary envelope.
      x-websocket-messages:
        - direction: server-to-client
          description: Per-frame inference envelope.
          schema:
            $ref: "#/components/schemas/StreamFrameEnvelope"
        - direction: server-to-client
          description: Session summary envelope sent when stream closes.
          schema:
            $ref: "#/components/schemas/StreamSummaryEnvelope"
        - direction: server-to-client
          description: Non-fatal processing error envelope.
          schema:
            $ref: "#/components/schemas/StreamErrorEnvelope"

  /v1/batches:
    post:
      operationId: createBatch
      summary: Create a harvest batch and trigger anchor workflow
      tags: [batch]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchCreateRequest"
      responses:
        "201":
          description: Batch created and anchored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
        "202":
          description: Batch accepted and saved, chain anchor pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
        "400":
          description: Invalid request or unripe confirmation missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden by policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicated batch conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Service unavailable or unrecoverable chain error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/batches/{batch_id}:
    get:
      operationId: getBatch
      summary: Get batch detail by batch id
      tags: [batch]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/BatchID"
      responses:
        "200":
          description: Batch detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden by policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Batch not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/trace/{trace_code}:
    get:
      operationId: getPublicTrace
      summary: Public trace query and verification result
      tags: [trace]
      security: []
      parameters:
        - $ref: "#/components/parameters/TraceCode"
      responses:
        "200":
          description: Public trace data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceResponse"
        "404":
          description: Trace code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Verify service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/dashboard/overview:
    get:
      operationId: getDashboardOverview
      summary: Get dashboard aggregate metrics
      tags: [dashboard]
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Dashboard overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardOverviewResponse"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden by policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/batches/reconcile:
    post:
      operationId: reconcileBatches
      summary: Trigger manual reconcile for pending anchor batches
      tags: [batch]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconcileRequest"
      responses:
        "202":
          description: Reconcile accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconcileResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden by policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No requested pending batch found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Reconcile service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer

  parameters:
    BatchID:
      name: batch_id
      in: path
      required: true
      schema:
        type: string
      description: Batch unique identifier.
    TraceCode:
      name: trace_code
      in: path
      required: true
      schema:
        type: string
      description: Public trace code.

  schemas:
    RipenessLabel:
      type: string
      enum: [green, half, red, young]

    HarvestSuggestion:
      type: string
      enum: [not_ready, partially_ready, ready, overripe_risk]

    BatchStatus:
      type: string
      enum: [pending_anchor, anchored, anchor_failed]

    VerifyStatus:
      type: string
      enum: [pass, fail, pending]

    UnripeHandling:
      type: string
      enum: [sorted_out]

    Detection:
      type: object
      required: [bbox, class_name, ripeness, confidence]
      properties:
        bbox:
          type: array
          items:
            type: number
          minItems: 4
          maxItems: 4
        class_name:
          type: string
          enum: [lychee]
        ripeness:
          $ref: "#/components/schemas/RipenessLabel"
        confidence:
          type: number
          minimum: 0
          maximum: 1
        track_id:
          type: integer
          nullable: true

    FrameSummary:
      type: object
      properties:
        total:
          type: integer
        green:
          type: integer
        half:
          type: integer
        red:
          type: integer
        young:
          type: integer

    FrameResult:
      type: object
      required: [frame_index, timestamp_ms, detections, frame_summary]
      properties:
        frame_index:
          type: integer
          minimum: 0
        timestamp_ms:
          type: integer
          minimum: 0
        detections:
          type: array
          items:
            $ref: "#/components/schemas/Detection"
        frame_summary:
          $ref: "#/components/schemas/FrameSummary"

    RipenessRatio:
      type: object
      properties:
        green:
          type: number
        half:
          type: number
        red:
          type: number
        young:
          type: number

    SessionSummary:
      type: object
      properties:
        total_detected:
          type: integer
        ripeness_ratio:
          $ref: "#/components/schemas/RipenessRatio"
        harvest_suggestion:
          $ref: "#/components/schemas/HarvestSuggestion"

    ModelMeta:
      type: object
      required: [model_version, schema_version, adapter, loaded]
      properties:
        model_version:
          type: string
        schema_version:
          type: string
        adapter:
          type: string
        loaded:
          type: boolean

    HealthResponse:
      type: object
      required: [status, model]
      properties:
        status:
          type: string
        model:
          $ref: "#/components/schemas/ModelMeta"

    CurrentModelResponse:
      type: object
      required: [model_version, schema_version, adapter, loaded]
      properties:
        model_version:
          type: string
        schema_version:
          type: string
        adapter:
          type: string
        loaded:
          type: boolean

    ImageInferResponse:
      type: object
      required: [model_version, schema_version, inference_ms, result]
      properties:
        model_version:
          type: string
        schema_version:
          type: string
        inference_ms:
          type: number
        result:
          $ref: "#/components/schemas/FrameResult"

    StreamFrameEnvelope:
      type: object
      required: [type, model_version, schema_version, result]
      properties:
        type:
          type: string
          enum: [frame]
        model_version:
          type: string
        schema_version:
          type: string
        result:
          $ref: "#/components/schemas/FrameResult"

    StreamSummaryEnvelope:
      type: object
      required: [type, model_version, schema_version, summary]
      properties:
        type:
          type: string
          enum: [summary]
        model_version:
          type: string
        schema_version:
          type: string
        summary:
          $ref: "#/components/schemas/SessionSummary"

    StreamErrorEnvelope:
      type: object
      required: [type, detail]
      properties:
        type:
          type: string
          enum: [error]
        detail:
          type: string

    GatewayHealthResponse:
      type: object
      required: [status, gateway]
      properties:
        status:
          type: string
          enum: [ok, degraded]
        gateway:
          type: string
        upstream:
          $ref: "#/components/schemas/HealthResponse"
        error:
          type: string

    BatchSummaryInput:
      type: object
      required: [total, green, half, red, young]
      properties:
        total:
          type: integer
          minimum: 0
        green:
          type: integer
          minimum: 0
        half:
          type: integer
          minimum: 0
        red:
          type: integer
          minimum: 0
        young:
          type: integer
          minimum: 0

    BatchSummary:
      type: object
      description: >
        Aggregated ripeness summary. unripe_count = green + young.
        unripe_ratio = unripe_count / total (total > 0). Default threshold is 0.15.
      required: [total, green, half, red, young, unripe_count, unripe_ratio, unripe_handling]
      properties:
        total:
          type: integer
          minimum: 0
        green:
          type: integer
          minimum: 0
        half:
          type: integer
          minimum: 0
        red:
          type: integer
          minimum: 0
        young:
          type: integer
          minimum: 0
        unripe_count:
          type: integer
          minimum: 0
        unripe_ratio:
          type: number
          minimum: 0
          maximum: 1
        unripe_handling:
          $ref: "#/components/schemas/UnripeHandling"

    BatchCreateRequest:
      type: object
      required: [orchard_id, orchard_name, plot_id, harvested_at, summary]
      properties:
        orchard_id:
          type: string
        orchard_name:
          type: string
        plot_id:
          type: string
        plot_name:
          type: string
        harvested_at:
          type: string
          format: date-time
        summary:
          $ref: "#/components/schemas/BatchSummaryInput"
        note:
          type: string
        confirm_unripe:
          type: boolean
          default: false
          description: >
            Required to be true when unripe_ratio > 0.15 in order to create the batch.

    AnchorProof:
      type: object
      required: [tx_hash, block_number, chain_id, contract_address, anchor_hash, anchored_at]
      properties:
        tx_hash:
          type: string
        block_number:
          type: integer
          minimum: 0
        chain_id:
          type: string
        contract_address:
          type: string
        anchor_hash:
          type: string
        anchored_at:
          type: string
          format: date-time

    Batch:
      type: object
      required:
        [batch_id, trace_code, status, orchard_id, orchard_name, plot_id, harvested_at, summary, created_at]
      properties:
        batch_id:
          type: string
        trace_code:
          type: string
        status:
          $ref: "#/components/schemas/BatchStatus"
        orchard_id:
          type: string
        orchard_name:
          type: string
        plot_id:
          type: string
        plot_name:
          type: string
          nullable: true
        harvested_at:
          type: string
          format: date-time
        summary:
          $ref: "#/components/schemas/BatchSummary"
        note:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        anchor_proof:
          oneOf:
            - $ref: "#/components/schemas/AnchorProof"
            - type: "null"

    TraceBatch:
      type: object
      required: [batch_id, trace_code, status, orchard_name, plot_name, harvested_at, summary, created_at]
      properties:
        batch_id:
          type: string
        trace_code:
          type: string
        status:
          $ref: "#/components/schemas/BatchStatus"
        orchard_name:
          type: string
        plot_name:
          type: string
        harvested_at:
          type: string
          format: date-time
        summary:
          $ref: "#/components/schemas/BatchSummary"
        created_at:
          type: string
          format: date-time

    TraceVerifyResult:
      type: object
      required: [verify_status, reason]
      properties:
        verify_status:
          $ref: "#/components/schemas/VerifyStatus"
        reason:
          type: string

    TraceResponse:
      type: object
      required: [batch, verify_result]
      properties:
        batch:
          $ref: "#/components/schemas/TraceBatch"
        verify_result:
          $ref: "#/components/schemas/TraceVerifyResult"

    DashboardTotals:
      type: object
      required: [batch_total]
      properties:
        batch_total:
          type: integer
          minimum: 0

    BatchStatusDistribution:
      type: object
      required: [anchored, pending_anchor, anchor_failed]
      properties:
        anchored:
          type: integer
          minimum: 0
        pending_anchor:
          type: integer
          minimum: 0
        anchor_failed:
          type: integer
          minimum: 0

    RipenessDistribution:
      type: object
      required: [green, half, red, young]
      properties:
        green:
          type: integer
          minimum: 0
        half:
          type: integer
          minimum: 0
        red:
          type: integer
          minimum: 0
        young:
          type: integer
          minimum: 0

    UnripeMetrics:
      type: object
      required: [unripe_batch_count, unripe_batch_ratio, threshold, unripe_handling]
      properties:
        unripe_batch_count:
          type: integer
          minimum: 0
        unripe_batch_ratio:
          type: number
          minimum: 0
          maximum: 1
        threshold:
          type: number
          minimum: 0
          maximum: 1
          default: 0.15
        unripe_handling:
          $ref: "#/components/schemas/UnripeHandling"

    RecentAnchorRecord:
      type: object
      required: [batch_id, trace_code, status, created_at]
      properties:
        batch_id:
          type: string
        trace_code:
          type: string
        status:
          $ref: "#/components/schemas/BatchStatus"
        tx_hash:
          type: string
          nullable: true
        anchored_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time

    ReconcileStats:
      type: object
      required: [pending_count, retried_total, failed_total]
      properties:
        pending_count:
          type: integer
          minimum: 0
        retried_total:
          type: integer
          minimum: 0
        failed_total:
          type: integer
          minimum: 0
        last_reconcile_at:
          type: string
          format: date-time
          nullable: true

    DashboardOverviewResponse:
      type: object
      required:
        [totals, status_distribution, ripeness_distribution, unripe_metrics, recent_anchors, reconcile_stats]
      properties:
        totals:
          $ref: "#/components/schemas/DashboardTotals"
        status_distribution:
          $ref: "#/components/schemas/BatchStatusDistribution"
        ripeness_distribution:
          $ref: "#/components/schemas/RipenessDistribution"
        unripe_metrics:
          $ref: "#/components/schemas/UnripeMetrics"
        recent_anchors:
          type: array
          items:
            $ref: "#/components/schemas/RecentAnchorRecord"
        reconcile_stats:
          $ref: "#/components/schemas/ReconcileStats"

    ReconcileRequest:
      type: object
      properties:
        batch_ids:
          type: array
          items:
            type: string
          description: >
            Optional explicit pending batch ids to reconcile.
            If omitted, the service selects pending batches by limit.
        limit:
          type: integer
          minimum: 1
          maximum: 1000
          default: 100

    ReconcileResponse:
      type: object
      required: [accepted, requested_count, scheduled_count, skipped_count]
      properties:
        accepted:
          type: boolean
        requested_count:
          type: integer
          minimum: 0
        scheduled_count:
          type: integer
          minimum: 0
        skipped_count:
          type: integer
          minimum: 0
        message:
          type: string

    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        request_id:
          type: string
        details:
          type: object
          additionalProperties: true